# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "GolfTracker API",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://backend-production-b6713.up.railway.app" },
    { "url": "http://localhost:3001" }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id":         { "type": "integer" },
          "email":      { "type": "string" },
          "password":   { "type": "string", "description": "MD5 hash" },
          "first_name": { "type": "string" },
          "last_name":  { "type": "string" },
          "handicap":   { "type": "number" },
          "role":       { "type": "string", "enum": ["user", "admin"] },
          "bio":        { "type": "string" },
          "avatar_url": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" }
        }
      },
      "Course": {
        "type": "object",
        "properties": {
          "id":            { "type": "integer" },
          "name":          { "type": "string" },
          "location":      { "type": "string" },
          "description":   { "type": "string" },
          "par":           { "type": "integer" },
          "slope_rating":  { "type": "number" },
          "course_rating": { "type": "number" },
          "holes":         { "type": "integer" },
          "created_at":    { "type": "string", "format": "date-time" }
        }
      },
...
      "AuthResponse": {
        "type": "object",
        "properties": {
          "token": { "type": "string" },
          "user":  { "$ref": "#/components/schemas/User" }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error":  { "type": "string" },
          "stack":  { "type": "string" },
          "detail": { "type": "string" },
          "query":  { "type": "string" },
          "hint":   { "type": "string" },
          "code":   { "type": "string" }
        }
      },
      "Message": {
        "type": "object",
        "properties": {
          "message": { "type": "string" }
        }
      }
    }
  },
  "paths": {
    "/api/auth/register": {
      "post": {
        "tags": ["Auth"],
        "summary": "Register a new user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password", "first_name", "last_name"],
                "properties": {
                  "email":      { "type": "string" },
                  "password":   { "type": "string" },
                  "first_name": { "type": "string" },
                  "last_name":  { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "User created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "500": { "description": "Error", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "summary": "Log in",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["email", "password"],
                "properties": {
                  "email":    { "type": "string" },
                  "password": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login successful", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/AuthResponse" } } } },
          "401": { "description": "Invalid credentials", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/courses": {
      "get": {
        "tags": ["Courses"],
        "summary": "List courses",
        "parameters": [
          { "name": "search",   "in": "query", "schema": { "type": "string" }, "description": "Filter by name" },
          { "name": "location", "in": "query", "schema": { "type": "string" }, "description": "Filter by location" }
        ],
        "responses": {
          "200": { "description": "Course list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Course" } } } } }
        }
      },
      "post": {
        "tags": ["Courses"],
        "summary": "Create a course",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "location", "description", "par", "slope_rating", "course_rating"],
                "properties": {
                  "name":          { "type": "string" },
                  "location":      { "type": "string" },
                  "description":   { "type": "string" },
                  "par":           { "type": "integer" },
                  "slope_rating":  { "type": "number" },
                  "course_rating": { "type": "number" },
                  "holes":         { "type": "integer", "default": 18 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created course", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } }
        }
      }
    },
    "/api/courses/fetch-image": {
      "get": {
        "tags": ["Courses"],
        "summary": "Fetch an image by URL (SSRF)",
        "parameters": [
          { "name": "url", "in": "query", "required": true, "schema": { "type": "string" }, "description": "URL to fetch" }
        ],
        "responses": {
          "200": { "description": "File content" },
          "400": { "description": "Missing url parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
    "/api/courses/{id}": {
      "get": {
        "tags": ["Courses"],
        "summary": "Get a course by ID",
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Course", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "put": {
        "tags": ["Courses"],
        "summary": "Update a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "location", "description", "par", "slope_rating", "course_rating"],
                "properties": {
                  "name":          { "type": "string" },
                  "location":      { "type": "string" },
                  "description":   { "type": "string" },
                  "par":           { "type": "integer" },
                  "slope_rating":  { "type": "number" },
                  "course_rating": { "type": "number" },
                  "holes":         { "type": "integer", "default": 18 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated course", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Courses"],
        "summary": "Delete a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } }
        }
      }
    },
...
    "/api/export/scorecard/{roundId}": {
      "get": {
        "tags": ["Export"],
        "summary": "Export a scorecard",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "roundId", "in": "path", "required": true, "schema": { "type": "integer" } },
          { "name": "format",  "in": "query", "schema": { "type": "string", "enum": ["json", "pdf", "csv"], "default": "json" } }
        ],
        "responses": {
          "200": { "description": "Scorecard data or export result" }
        }
      }
    },
    "/api/export/file": {
      "get": {
        "tags": ["Export"],
        "summary": "Download an export file by name (path traversal)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "name", "in": "query", "required": true, "schema": { "type": "string" }, "description": "Filename in the exports directory" }
        ],
        "responses": {
          "200": { "description": "File content" },
          "400": { "description": "Missing name parameter", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      }
    },
...
    "/api/admin/courses": {
      "get": {
        "tags": ["Admin"],
        "summary": "List all courses",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": { "description": "Course list", "content": { "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Course" } } } } }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Create a course",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "location", "description", "par", "slope_rating", "course_rating"],
                "properties": {
                  "name":          { "type": "string" },
                  "location":      { "type": "string" },
                  "description":   { "type": "string" },
                  "par":           { "type": "integer" },
                  "slope_rating":  { "type": "number" },
                  "course_rating": { "type": "number" },
                  "holes":         { "type": "integer", "default": 18 }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Created course", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } }
        }
      }
    },
    "/api/admin/courses/{id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Update a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "location", "description", "par", "slope_rating", "course_rating"],
                "properties": {
                  "name":          { "type": "string" },
                  "location":      { "type": "string" },
                  "description":   { "type": "string" },
                  "par":           { "type": "integer" },
                  "slope_rating":  { "type": "number" },
                  "course_rating": { "type": "number" },
                  "holes":         { "type": "integer", "default": 18 }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Updated course", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Course" } } } },
          "404": { "description": "Not found", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Delete a course",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
        "responses": {
          "200": { "description": "Deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Message" } } } }
        }
      }
    },
...
  }
}